(PC-20410) refacto(index): move types from index to algolia types + finally delete libs/search folder
