+++ src/features/home/components/modules/venues/VenueTile.tsx
@@ -7,12 +7,12 @@ import styled from 'styled-components/native'
 import { VenueResponse } from 'api/gen'
 import { VenueDetails } from 'features/home/components/modules/venues/VenueDetails'
 import { VenueTypeLocationIcon } from 'features/home/components/VenueTypeLocationIcon'
+import { VenueHit } from 'libs/algolia'
 import { analytics } from 'libs/firebase/analytics'
 import { GeoCoordinates } from 'libs/geolocation'
 import { useHandleFocus } from 'libs/hooks/useHandleFocus'
 import { formatDistance, mapVenueTypeToIcon } from 'libs/parsers'
 import { QueryKeys } from 'libs/queryKeys'
-import { VenueHit } from 'libs/search'
 import { tileAccessibilityLabel, TileContentType } from 'libs/tileAccessibilityLabel'
 import { ImageTile } from 'ui/components/ImageTile'
 import { InternalTouchableLink } from 'ui/components/touchableLink/InternalTouchableLink'
