+++ src/features/home/components/modules/venues/VenuesModule.tsx
@@ -2,10 +2,10 @@ import React, { useCallback, useEffect } from 'react'
 
 import { useVenueModule } from 'features/home/api/useVenueModule'
 import { VenueTile } from 'features/home/components/modules/venues/VenueTile'
+import { VenueHit } from 'libs/algolia'
 import { ContentTypes, DisplayParametersFields, VenuesParametersFields } from 'libs/contentful'
 import { analytics } from 'libs/firebase/analytics'
 import { useGeolocation } from 'libs/geolocation'
-import { VenueHit } from 'libs/search'
 import { PassPlaylist } from 'ui/components/PassPlaylist'
 import { CustomListRenderItem } from 'ui/components/Playlist'
 import { LENGTH_S } from 'ui/theme'
