+++ src/features/offer/pages/Offer/Offer.native.test.tsx
@@ -5,10 +5,10 @@ import { BookingsResponse } from 'api/gen'
 import { useAuthContext } from 'features/auth/context/AuthContext'
 import { offerId, renderOfferPage } from 'features/offer/helpers/renderOfferPageTestUtil'
 import { beneficiaryUser } from 'fixtures/user'
+import { SearchHit } from 'libs/algolia'
 import { mockedAlgoliaResponse } from 'libs/algolia/__mocks__/mockedAlgoliaResponse'
 import { env } from 'libs/environment'
 import { analytics } from 'libs/firebase/analytics'
-import { SearchHit } from 'libs/search'
 import { server } from 'tests/server'
 import { act, fireEvent, screen, waitFor } from 'tests/utils'
