+++ src/features/search/components/SearchResults/SearchResults.native.test.tsx
@@ -10,11 +10,11 @@ import { LocationType } from 'features/search/enums'
 import { MAX_RADIUS } from 'features/search/helpers/reducer.helpers'
 import { LocationFilter, SearchState, UserData } from 'features/search/types'
 import { beneficiaryUser, nonBeneficiaryUser } from 'fixtures/user'
+import { SearchHit } from 'libs/algolia'
 import { mockedAlgoliaResponse } from 'libs/algolia/__mocks__/mockedAlgoliaResponse'
 import { analytics } from 'libs/firebase/analytics'
 import { GeoCoordinates } from 'libs/geolocation'
 import { SuggestedPlace } from 'libs/place'
-import { SearchHit } from 'libs/search'
 import { placeholderData as mockSubcategoriesData } from 'libs/subcategories/placeholderData'
 import { SuggestedVenue } from 'libs/venue'
 import { mockedSuggestedVenues } from 'libs/venue/fixtures/mockedSuggestedVenues'
