+++ src/features/search/components/SearchResults/SearchResults.tsx
@@ -29,11 +29,11 @@ import { LocationModal } from 'features/search/pages/modals/LocationModal/Locati
 import { OfferDuoModal } from 'features/search/pages/modals/OfferDuoModal/OfferDuoModal'
 import { PriceModal } from 'features/search/pages/modals/PriceModal/PriceModal'
 import { AccessibilityRole } from 'libs/accessibilityRole/accessibilityRole'
+import { SearchHit } from 'libs/algolia'
 import { analytics } from 'libs/firebase/analytics'
 import { useGeolocation } from 'libs/geolocation'
 import { useIsFalseWithDelay } from 'libs/hooks/useIsFalseWithDelay'
 import { plural } from 'libs/plural'
-import { SearchHit } from 'libs/search'
 import { useOpacityTransition } from 'ui/animations/helpers/useOpacityTransition'
 import { Banner } from 'ui/components/Banner'
 import { ButtonSecondary } from 'ui/components/buttons/ButtonSecondary'
