+++ src/features/venue/components/VenueOffers/VenueOffers.native.test.tsx
@@ -9,8 +9,8 @@ import { useVenueOffers } from 'features/venue/api/useVenueOffers'
 import { VenueOffers } from 'features/venue/components/VenueOffers/VenueOffers'
 import { VenueOffersResponseSnap } from 'features/venue/fixtures/venueOffersResponseSnap'
 import { venueResponseSnap } from 'features/venue/fixtures/venueResponseSnap'
+import { SearchHit } from 'libs/algolia'
 import { analytics } from 'libs/firebase/analytics'
-import { SearchHit } from 'libs/search'
 import { placeholderData } from 'libs/subcategories/placeholderData'
 import { fireEvent, render } from 'tests/utils'
