+++ src/libs/algolia/fetchAlgolia/fetchMultipleOffers/fetchMultipleOffers.ts
@@ -1,6 +1,7 @@
 import flatten from 'lodash/flatten'
 
 import { SearchState } from 'features/search/types'
+import { SearchHit } from 'libs/algolia'
 import { captureAlgoliaError } from 'libs/algolia/fetchAlgolia/AlgoliaError'
 import { buildOfferSearchParameters } from 'libs/algolia/fetchAlgolia/buildAlgoliaParameters/buildOfferSearchParameters.ts'
 import { offerAttributesToRetrieve } from 'libs/algolia/fetchAlgolia/buildAlgoliaParameters/offerAttributesToRetrieve'
@@ -8,7 +9,6 @@ import { client } from 'libs/algolia/fetchAlgolia/clients'
 import { buildHitsPerPage } from 'libs/algolia/fetchAlgolia/utils'
 import { env } from 'libs/environment'
 import { GeoCoordinates } from 'libs/geolocation'
-import { SearchHit } from 'libs/search'
 
 type FetchMultipleOffersArgs = {
   paramsList: SearchState[]
