+++ src/libs/parsers/formatDistance.ts
@@ -1,5 +1,5 @@
+import { Geoloc } from 'libs/algolia'
 import { GeoCoordinates } from 'libs/geolocation'
-import { Geoloc } from 'libs/search'
 
 const EARTH_RADIUS_KM = 6378.137
